create gateway just for case status, return domain objects to usecase
