+++ SocialCareCaseViewerApi/V1/Gateways/DatabaseGateway.cs
@@ -1010,92 +1010,6 @@ namespace SocialCareCaseViewerApi.V1.Gateways
             _databaseContext.SaveChanges();
         }
 
-        public IEnumerable<Infrastructure.CaseStatus> GetCaseStatusesByPersonId(long personId)
-        {
-            var caseStatuses = _databaseContext.CaseStatuses.Where(cs => cs.PersonId == personId)
-                .Where(cs => cs.EndDate == null || cs.EndDate > DateTime.Today)
-                .Include(cs => cs.Type)
-                .Include(cs => cs.SelectedOptions)
-                .ThenInclude(csso => csso.FieldOption)
-                .ThenInclude(fo => fo.TypeField);
-
-            return caseStatuses;
-        }
-        public Infrastructure.CaseStatus GetCaseStatusesByPersonIdDate(long personId, DateTime date)
-        {
-            var caseStatuse = _databaseContext.CaseStatuses.Where(cs => cs.PersonId == personId)
-                .Where(cs => cs.StartDate <= date)
-                .Where(cs => cs.EndDate == null || cs.EndDate >= date)
-                .Include(cs => cs.Type)
-                .Include(cs => cs.SelectedOptions)
-                .ThenInclude(csso => csso.FieldOption)
-                .ThenInclude(fo => fo.TypeField)
-                .FirstOrDefault();
-
-            return caseStatuse;
-        }
-
-        public Infrastructure.CaseStatusType GetCaseStatusTypeWithFields(string type)
-        {
-            var response = _databaseContext.CaseStatusTypes
-                .Where(cs => cs.Name == type)
-                .Include(cs => cs.Fields)
-                .ThenInclude(sf => sf.Options);
-
-            return response.FirstOrDefault();
-        }
-
-        public Infrastructure.CaseStatus CreateCaseStatus(CreateCaseStatusRequest request)
-        {
-
-            var statusType = _databaseContext.CaseStatusTypes
-                .Where(f => f.Name == request.Type)
-                .FirstOrDefault();
-
-            var caseStatus = new Infrastructure.CaseStatus()
-            {
-                PersonId = request.PersonId,
-                TypeId = statusType.Id,
-                StartDate = request.StartDate,
-                Notes = request.Notes,
-                CreatedBy = request.CreatedBy
-            };
-
-            _databaseContext.CaseStatuses.Add(caseStatus);
-
-            if (request.Fields != null)
-            {
-                foreach (var optionValue in request.Fields)
-                {
-                    var field = _databaseContext.CaseStatusTypeFields
-                        .FirstOrDefault(f => f.Name == optionValue.Name);
-
-                    var fieldTypeOption = _databaseContext.CaseStatusTypeFieldOptions
-                        .Where(fov => fov.Name == optionValue.Selected)
-                        .FirstOrDefault(fov => fov.TypeFieldId == field.Id);
-
-                    if (fieldTypeOption != null)
-                    {
-                        var fieldOption = new CaseStatusFieldOption
-                        {
-                            StatusId = caseStatus.Id,
-                            FieldOptionId = fieldTypeOption.Id
-                        };
-
-                        if (caseStatus.SelectedOptions == null)
-                        {
-                            caseStatus.SelectedOptions = new List<CaseStatusFieldOption>();
-                        }
-                        caseStatus.SelectedOptions.Add(fieldOption);
-                    }
-                }
-            }
-
-            _databaseContext.SaveChanges();
-
-            return caseStatus;
-        }
-
         private static AllocationSet SetDeallocationValues(AllocationSet allocation, DateTime dt, string modifiedBy)
         {
             //keep workerId and TeamId in the record so they can be easily exposed to front end
