+++ SocialCareCaseViewerApi/V1/Gateways/ICaseStatusGateway.cs
@@ -1,0 +1,14 @@
+using System.Collections.Generic;
+using SocialCareCaseViewerApi.V1.Boundary.Requests;
+using SocialCareCaseViewerApi.V1.Domain;
+
+namespace SocialCareCaseViewerApi.V1.Gateways
+{
+    public interface ICaseStatusGateway
+    {
+        IEnumerable<CaseStatus> GetCaseStatusesByPersonId(long personId);
+        CaseStatus GetCaseStatusesByPersonIdDate(long personId, System.DateTime period);
+        Infrastructure.CaseStatusType GetCaseStatusTypeWithFields(string caseStatusType);
+        CaseStatus CreateCaseStatus(CreateCaseStatusRequest request);
+    }
+}
