+++ SocialCareCaseViewerApi/V1/Gateways/IDatabaseGateway.cs
@@ -41,10 +41,6 @@ namespace SocialCareCaseViewerApi.V1.Gateways
         PersonalRelationship CreatePersonalRelationship(CreatePersonalRelationshipRequest request);
         PersonalRelationship GetPersonalRelationshipById(long relationshipId);
         void DeleteRelationship(long relationshipId);
-        IEnumerable<CaseStatus> GetCaseStatusesByPersonId(long personId);
-        CaseStatus GetCaseStatusesByPersonIdDate(long personId, System.DateTime period);
-        CaseStatusType GetCaseStatusTypeWithFields(string caseStatusType);
-        CaseStatus CreateCaseStatus(CreateCaseStatusRequest request);
         void CreateRequestAudit(CreateRequestAuditRequest request);
     }
 }
