+++ SocialCareCaseViewerApi/V1/UseCase/Interfaces/ICaseStatusesUseCase.cs
@@ -1,6 +1,6 @@
 using SocialCareCaseViewerApi.V1.Boundary.Requests;
 using SocialCareCaseViewerApi.V1.Boundary.Response;
-using SocialCareCaseViewerApi.V1.Infrastructure;
+using SocialCareCaseViewerApi.V1.Domain;
 
 namespace SocialCareCaseViewerApi.V1.UseCase.Interfaces
 {
